backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
pgvector==0.3.3
python-multipart==0.0.9
pypdf==5.0.1
python-docx==1.1.2
sentence-transformers==3.0.1
numpy==1.26.4
httpx==0.27.2
